requirements.txt
django
django-rest-framework
django-extensions
django-environ
django-filter
django-redis
psycopg2-binary
pandas
gunicorn